Add a Subscribe() function to the Consensus interface
In order to build consensus systems it is necessary that every consensus member
can enjoy the possibility of subscribing to any updates in the state.
Otherwise they must GetCurrentState() repeteadly and innefficiently.

I have opted for a channel to do this because they seem appropiate for the use
case, although a similar behaviour would be achieved with something like
SetNewStateCallback(func(State)).

License: MIT
Signed-off-by: Hector Sanjuan <hector@protocol.ai>